+++ src/main/kotlin/backend/itracker/ItrackerApplication.kt
@@ -1,6 +1,6 @@
 package backend.itracker
 
-import backend.itracker.schedule.config.AuditConfig
+import backend.itracker.config.AuditConfig
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.boot.runApplication
