+++ src/main/kotlin/backend/itracker/schedule/service/crawlers/Schedulable.kt
@@ -1,0 +1,16 @@
+package backend.itracker.schedule.service.crawlers
+
+import backend.itracker.crawl.service.CrawlResultService
+import backend.itracker.crawl.service.CrawlService
+import backend.itracker.crawl.service.CrawlTargetCategory
+
+abstract class Schedulable(
+    private val crawlResultService: CrawlResultService,
+    private val crawlService: CrawlService,
+) {
+
+    abstract fun supports(): CrawlTargetCategory
+
+    abstract fun doSchedule()
+}
+
